Output param name in param comment text.
The text displayed in <param name="xxx">The text.</param> was originally hardcoded to "The text." to satisfy the unit test in the previous commit. This has been changed to output the name of the parameter instead of "text".

The unit test was changed from a [Fact] to a [Theory] to allow testing multiple parameter names.