Create param text for param name that ends in "name".
Splits camel case parameter names into separate words. Convert each word into lower case. If last word is "name", then output "Name of the ..." as the parameter text. Otherwise, output "The ..." as the parameter text.

There are a number of unintended consequences in this commit. For example, a parameter name such as "theIBMComputer" is displayed "The the ibm computer". These will be fixed in future commits.