Modify name param text for one of first words.
Up to this point, any parameter name that ends in "Name" has the parameter description that begins "Name of the". This change modifies the description text for any parameter name that begins with one of the previously defined first words. For example, a parameter name of "aSyntaxNodeName" will have the text description of "Name of a syntax node." instead of "Name of the a syntax node."